backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Pillow>=10.0.0
numpy>=1.26.0
opencv-python>=4.8.0
Werkzeug>=3.0.0

